Fix html attribute
In order to know if the completion must be triggered, the code must know when the cursor is inside a class attribut in an html file. This commit fix the algorithm used to detect html attributes. The initial algorithm expected that attributes are separated by spaces, but they can be separated by other whitespace characters (such as tabulations). 

Note that, funnily enough, the same algorithm is used by the sublime angular plugin, and I proposed the same change here: https://github.com/angular-ui/AngularJS-sublime-package/pull/72/files